refactor(partners): consolidar specs de dashboard en partners.service.spec y partners.controller.spec
- Fusiona dashboard.service.spec.ts en partners.service.spec.ts (getPartnerMetrics, getPayments)
- Renombra dashboard.controller.spec.ts a partners.controller.spec.ts
- Elimina archivos de spec redundantes
